// src/app/core/api/generated/model/ingredientPriceSourceResponse.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const IngredientPriceSourceResponse = {
    Organization: 'organization',
    Store: 'store'
} as const;
export type IngredientPriceSourceResponse = typeof IngredientPriceSourceResponse[keyof typeof IngredientPriceSourceResponse];