// src/app/core/api/generated/model/managedIngredientResponse.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ManagedIngredientHistoryPriceResponse } from './managedIngredientHistoryPriceResponse';
import { IngredientCategorySummaryResponse } from './ingredientCategorySummaryResponse';


/**
 * Resposta agregada da gestão detalhada de um ingrediente.
 */
export interface ManagedIngredientResponse { 
    /**
     * Identificador do ingrediente.
     */
    id?: string;
    /**
     * Nome canônico do ingrediente.
     */
    name?: string | null;
    /**
     * Unidade base estrutural do ingrediente.
     */
    baseUnit?: string | null;
    category?: IngredientCategorySummaryResponse;
    /**
     * Histórico agregado de preços elegíveis para a consulta.
     */
    historyPrices?: Array<ManagedIngredientHistoryPriceResponse> | null;
}